+++ src/actions/names.js
@@ -15,6 +15,7 @@ export const ACTIONS = Object.freeze({
     setSection: 'ui.setSection',
     toggleOverlay: 'ui.toggleOverlay',
     setTheme: 'ui.setTheme',
+    svgAttr: 'ui.svgAttr',
     setCandleInterval: 'ui.setCandleInterval',
     palette: 'ui.palette',
     paletteSearch: 'ui.paletteSearch',